chore: re-organize faucet code
