+++ src/faucet/mod.rs
@@ -1,0 +1,3 @@
+pub mod views;
+pub mod utils;
+
